Implement opti on ios

+++ ios/Classes/SwiftTtsPlugin.swift
@@ -25,20 +25,20 @@ public class SwiftTtsPlugin: NSObject, FlutterPlugin {
             let argurments = call.arguments as! Dictionary<String, Any>
             let fastSpeechModel = argurments["fastSpeechModel"] as! String
             let melganModel = argurments["melganModel"] as! String
-            let inputIds = argurments["inputIds"] as! Array<Int32>
+            let inputIds = argurments["inputIds"] as! Array<Int64>
             let speed = argurments["speed"] as! NSNumber
             let speakerId = argurments["speakerId"] as! Int
             let sampleRate = argurments["sampleRate"] as! Int
             let hopSize = argurments["hopSize"] as! Int
             tts.logEnabled = argurments["logEnabled"] as? Bool ?? true
             tts.threadCount = argurments["threadCount"] as? Int ?? 1
-            tts.speak(fastSpeechModel: fastSpeechModel, melGanModel: melganModel, inputIds: inputIds, speakerId: Int32(speakerId), speed: Float(truncating: speed), sampleRate: sampleRate, hopSize: hopSize, result: result)
+            tts.speak(fastSpeechModel: fastSpeechModel, melGanModel: melganModel, inputIds: inputIds, speakerId: Int64(speakerId), speed: Float(truncating: speed), sampleRate: sampleRate, hopSize: hopSize, result: result)
             break
           case "generateVoice":
             let argurments = call.arguments as! Dictionary<String, Any>
             let fastSpeechModel = argurments["fastSpeechModel"] as! String
             let melganModel = argurments["melganModel"] as! String
-            let inputIds = argurments["inputIds"] as! Array<Int32>
+            let inputIds = argurments["inputIds"] as! Array<Int64>
             let speed = argurments["speed"] as! NSNumber
             let speakerId = argurments["speakerId"] as! Int
             let sampleRate = argurments["sampleRate"] as! Int
@@ -47,13 +47,13 @@ public class SwiftTtsPlugin: NSObject, FlutterPlugin {
             let singleThread = argurments["singleThread"] as! Bool
             tts.logEnabled = argurments["logEnabled"] as? Bool ?? true
             tts.threadCount = argurments["threadCount"] as? Int ?? 1
-            tts.generateVoice(requestId: requestId, fastSpeechModel: fastSpeechModel, melGanModel: melganModel, inputIds: inputIds, speakerId: Int32(speakerId), speed: Float(truncating: speed), sampleRate: sampleRate, hopSize: hopSize, singleThread:singleThread, result: result)
+            tts.generateVoice(requestId: requestId, fastSpeechModel: fastSpeechModel, melGanModel: melganModel, inputIds: inputIds, speakerId: Int64(speakerId), speed: Float(truncating: speed), sampleRate: sampleRate, hopSize: hopSize, singleThread:singleThread, result: result)
             break
           case "playVoice":
             let argurments = call.arguments as! Dictionary<String, Any>
             let fastSpeechModel = argurments["fastSpeechModel"] as! String
             let melganModel = argurments["melganModel"] as! String
-            let inputIds = argurments["inputIds"] as! Array<Int32>
+            let inputIds = argurments["inputIds"] as! Array<Int64>
             let speed = argurments["speed"] as! NSNumber
             let speakerId = argurments["speakerId"] as! Int
             let sampleRate = argurments["sampleRate"] as! Int
@@ -63,7 +63,7 @@ public class SwiftTtsPlugin: NSObject, FlutterPlugin {
             let playerCompletedDelayed = argurments["playerCompletedDelayed"] as! Int
             tts.logEnabled = argurments["logEnabled"] as? Bool ?? true
             tts.threadCount = argurments["threadCount"] as? Int ?? 1
-            tts.playVoice(requestId: requestId, fastSpeechModel: fastSpeechModel, melGanModel: melganModel, inputIds: inputIds, speakerId: Int32(speakerId), speed: Float(truncating: speed), sampleRate: sampleRate, hopSize: hopSize, singleThread:singleThread, playerCompletedDelayed: playerCompletedDelayed, result: result)
+            tts.playVoice(requestId: requestId, fastSpeechModel: fastSpeechModel, melGanModel: melganModel, inputIds: inputIds, speakerId: Int64(speakerId), speed: Float(truncating: speed), sampleRate: sampleRate, hopSize: hopSize, singleThread:singleThread, playerCompletedDelayed: playerCompletedDelayed, result: result)
             break
             case "dispose":
             tts.dispose()
